+++ RELEASE_NOTES.md
@@ -1,5 +1,7 @@
 ### New in 0.4
 
+* Braking: `ValueObject` now uses public properties instead of both
+  private and public fields
 * Feature: Release notes added to NuGet packages
 * Fixed: Unchecked missing in `ValueObject` when claculating hash
 * Fixed: `NullReferenceException` thrown if `null` was stored